Update docs and skip copy of dummy sky models

+++ rapthor/operations/image.py
@@ -268,15 +268,19 @@ class Image(Operation):
 
             # The output sky models, both true sky and apparent sky (the filenames are
             # defined in the rapthor/scripts/filter_skymodel.py file)
-            sector.image_skymodel_file_true_sky = image_root + '.true_sky.txt'
-            sector.image_skymodel_file_apparent_sky = image_root + '.apparent_sky.txt'
-            dst_dir = os.path.join(self.parset['dir_working'], 'skymodels', 'image_{}'.format(self.index))
-            misc.create_directory(dst_dir)
-            for src_filename in [sector.image_skymodel_file_true_sky, sector.image_skymodel_file_apparent_sky]:
-                dst_filename = os.path.join(dst_dir, os.path.basename(src_filename))
-                if os.path.exists(dst_filename):
-                    os.remove(dst_filename)
-                shutil.copy(src_filename, dst_filename)
+            #
+            # Note: these are not generated when QUV images are made (WSClean does not
+            # currently support writing a source list in this mode)
+            if self.field.image_pol.lower() == 'i':
+                sector.image_skymodel_file_true_sky = image_root + '.true_sky.txt'
+                sector.image_skymodel_file_apparent_sky = image_root + '.apparent_sky.txt'
+                dst_dir = os.path.join(self.parset['dir_working'], 'skymodels', 'image_{}'.format(self.index))
+                misc.create_directory(dst_dir)
+                for src_filename in [sector.image_skymodel_file_true_sky, sector.image_skymodel_file_apparent_sky]:
+                    dst_filename = os.path.join(dst_dir, os.path.basename(src_filename))
+                    if os.path.exists(dst_filename):
+                        os.remove(dst_filename)
+                    shutil.copy(src_filename, dst_filename)
 
             # The output ds9 region file, if made
             if self.field.dde_method == 'facets':
